Split heat_mitigation.rs (780 lines) into submodule directory
Break the monolithic heat_mitigation.rs into focused submodules:
- constants.rs (32 lines): all mitigation-related constants
- state.rs (126 lines): HeatMitigationState resource + Saveable impl
- calculations.rs (78 lines): pure helper functions for effects/costs
- systems.rs (100 lines): Bevy system + plugin registration
- tests.rs (424 lines): all unit tests
- mod.rs (32 lines): module declarations + re-exports

All pub items are re-exported from mod.rs for full backwards
compatibility. No functional changes.

Part of #1389
